requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    "make update-requirements"
#
alabaster==0.7.12
    # via sphinx
arrow==1.2.0
    # via jinja2-time
astroid==2.8.0
    # via pylint
attrs==21.2.0
    # via
    #   pytest
    #   syrupy
babel==2.9.1
    # via sphinx
backports.entry-points-selectable==1.1.0
    # via virtualenv
binaryornot==0.4.4
    # via cookiecutter
black==21.9b0
    # via -r requirements-dev.in
bugyi.lib==0.6.1
    # via
    #   -r requirements.in
    #   pylogutils
    #   python-clap
bump2version==1.0.1
    # via -r requirements-dev.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.6
    # via requests
click==7.1.2
    # via
    #   black
    #   cookiecutter
    #   cruft
    #   pip-tools
    #   typer
colorama==0.4.4
    # via pylogutils
colored==1.4.2
    # via syrupy
cookiecutter==1.7.3
    # via cruft
coverage[toml]==6.0
    # via pytest-cov
cruft[pyproject]==2.9.0
    # via -r requirements-dev.in
distlib==0.3.3
    # via virtualenv
docutils==0.17.1
    # via
    #   m2r2
    #   sphinx
    #   sphinx-rtd-theme
filelock==3.3.0
    # via
    #   tox
    #   virtualenv
flake8==3.9.2
    # via -r requirements-dev.in
gitdb==4.0.7
    # via gitpython
gitpython==3.1.24
    # via cruft
idna==3.2
    # via requests
imagesize==1.2.0
    # via sphinx
importlib-metadata==4.8.1
    # via
    #   backports.entry-points-selectable
    #   flake8
    #   pep517
    #   pluggy
    #   pytest
    #   python-clap
    #   tox
    #   virtualenv
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via
    #   -r requirements-dev.in
    #   pylint
jinja2==3.0.2
    # via
    #   cookiecutter
    #   jinja2-time
    #   sphinx
jinja2-time==0.2.0
    # via cookiecutter
lazy-object-proxy==1.6.0
    # via astroid
m2r2==0.3.1
    # via -r requirements-dev.in
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mistune==0.8.4
    # via m2r2
mypy==0.910
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.0
    # via
    #   pytest
    #   setuptools-scm
    #   sphinx
    #   tox
pathspec==0.9.0
    # via black
pep517==0.11.0
    # via pip-tools
pip-tools==6.3.0
    # via -r requirements-dev.in
platformdirs==2.4.0
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via
    #   -r requirements.in
    #   pytest
    #   tox
poyo==0.5.0
    # via cookiecutter
py==1.10.0
    # via
    #   pytest
    #   tox
pycodestyle==2.7.0
    # via flake8
pydantic==1.8.2
    # via
    #   -r requirements.in
    #   pylogutils
    #   python-clap
pydocstyle[toml]==6.1.1
    # via -r requirements-dev.in
pyflakes==2.3.1
    # via flake8
pygments==2.10.0
    # via sphinx
pylint==2.11.1
    # via -r requirements-dev.in
pylogutils==0.1.4
    # via
    #   -r requirements.in
    #   python-clap
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-mock
    #   pytest-subprocess
    #   syrupy
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-mock==3.6.1
    # via -r requirements-dev.in
pytest-subprocess==1.1.2
    # via -r requirements-dev.in
python-clap==0.2.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   bugyi.lib
python-slugify==5.0.2
    # via cookiecutter
pytz==2021.3
    # via babel
regex==2021.9.30
    # via black
requests==2.26.0
    # via
    #   cookiecutter
    #   sphinx
setuptools-scm==6.3.2
    # via -r requirements-dev.in
six==1.16.0
    # via
    #   cookiecutter
    #   python-dateutil
    #   tox
    #   virtualenv
smmap==4.0.0
    # via gitdb
snowballstemmer==2.1.0
    # via
    #   pydocstyle
    #   sphinx
sphinx==4.2.0
    # via
    #   -r requirements-dev.in
    #   sphinx-autodoc-typehints
    #   sphinx-rtd-theme
sphinx-autodoc-typehints==1.12.0
    # via -r requirements-dev.in
sphinx-rtd-theme==1.0.0
    # via -r requirements-dev.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
structlog==21.1.0
    # via pylogutils
syrupy==1.4.5
    # via -r requirements-dev.in
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   cruft
    #   mypy
    #   pydocstyle
    #   pylint
    #   pytest
    #   tox
tomli==1.2.1
    # via
    #   black
    #   coverage
    #   pep517
    #   setuptools-scm
tox==3.24.4
    # via
    #   -r requirements-dev.in
    #   tox-pyenv
tox-pyenv==1.1.0
    # via -r requirements-dev.in
typed-ast==1.4.3
    # via
    #   astroid
    #   black
    #   mypy
typeguard==2.12.1
    # via -r requirements-dev.in
typer==0.3.2
    # via cruft
typing-extensions==3.10.0.2
    # via
    #   arrow
    #   astroid
    #   black
    #   bugyi.lib
    #   gitpython
    #   importlib-metadata
    #   mypy
    #   pydantic
    #   pylint
    #   structlog
urllib3==1.26.7
    # via requests
virtualenv==20.8.1
    # via tox
wheel==0.37.0
    # via pip-tools
wrapt==1.12.1
    # via astroid
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
pip==21.2.4
    # via pip-tools
setuptools==58.2.0
    # via
    #   astroid
    #   pip-tools
    #   setuptools-scm
    #   sphinx
